backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.4,<6.0
django-cors-headers>=4.6,<5.0
psycopg[binary]>=3.1
gunicorn>=23.0
requests>=2.32
dj-database-url>=2.3
Pillow>=11.0
anthropic>=0.40
